cmd: move completion command to its own subdirectory
While here, remove the need to provide an `io.Writer` and relay on the
one output set via `cobra.
